exercises_solution/tests/requirements.txt
pytest==8.3.2
python-dotenv
pytest-timeout==2.1.0
boto3==1.35.8
allure-pytest==2.13.5
great-expectations==1.1.0
great-expectations[s3]
# great_expectations[spark]
great-expectations[pandas]
pyarrow